rename module: bls_api -> bls
