+++ py_ecc/bls/__init__.py
@@ -1,0 +1,8 @@
+from .api import (  # noqa: F401
+    aggregate_pubkeys,
+    aggregate_signatures,
+    privtopub,
+    sign,
+    verify,
+    verify_multiple,
+)
